README.md
This is a creative software project based on sensor capabilities of any smartphone
running on Android. Android accounts for approximately 85% of all devices sold
today hence the famous Android platform is offering free application development
courses which can be used in various fields. In this study case the following sensors
served as an essential background knowledge: accelerometer, gyroscope and
magnetometer. The main idea stands behind developing an Android application
running on any Android device and collect real time sensor data for an indefinite
period while moving it. The needed materials for achieving this project are the
following: mastering Java programming language, an adequate usage of Android
Studio software and understanding smartphone sensor features. These were covered
by completing online modules regarding these topics followed by researching
additional books. Therefore, an Android app was designed in such a way that the
user can select any combination of the previous mentioned sensors in order to
display real-time graphs which can identify the evolution of the sensors’ stability
over time. Another feature is to store the modifying sensor values inside a csv file
which is useful for monitoring device movement such as tilt, shake, rotation, or
swing. Various tests with respect to sensors motion can be made wherefrom the
data generated by this app as a csv file saved for every recording in the internal
memory of the device is meant to be used in research considering a more in-depth
analysis with the help of other engineering tools such MATLAB. Ultimately, an
Android application that can collect sensor smartphone information was successfully
delivered in this project.